requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
passlib==1.7.4
pytz==2023.3
requests==2.31.0
speaklater==1.3
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
psycopg2-binary
